+++ lib/flatplan/presenter.rb
@@ -1,4 +1,5 @@
 require_relative 'presenter/manifest_serializer'
+require_relative 'presenter/console_flatplan'
 require_relative 'presenter/pandoc_manifest_serializer'
 
 module Flatplan
@@ -10,4 +11,5 @@ module Flatplan
   # Internal shortcut for rapid manifest transformation within services
   SerializeManifest = Presenter::ManifestSerializer
   SerializePandocManifest = Presenter::PandocManifestSerializer
+  RenderConsoleFlatplan = Presenter::ConsoleFlatplan
 end