+++ test/flatplan/dummy.rb
@@ -103,4 +103,4 @@ module Dummy
   end
 end
 
-pp Dummy.media_assets
+# pp Dummy.media_assets