+++ test/flatplan/service/test_dry_run.rb
@@ -6,7 +6,6 @@ describe 'Manifest Serialize/Load dry-run' do
     Dummy.publications.each do |origin|
       content = Presenter::ManifestSerializer.call(origin)
       parsed  = Builder::ManifestParser.call(content:)
-      pp origin, parsed
       assert_equal origin, parsed      
     end
   end